refactor(models, dao): Remove outdated `Branch` model and DAO, implement enriched `BranchDao`
- Deleted `Branch` class and its dependencies.
- Upgraded `BranchDao` to align with enhanced database schema, including warehouse counts and property details.
- Revised connection method in `BranchDao` for improved reliability.
- Updated references in `SuperAdminController` for consistency with the new implementation.
