+++ src/main/java/mx/edu/utez/warehousemanagerfx/Models/BranchDao.java
@@ -1,6 +1,5 @@
-package mx.edu.utez.warehousemanagerfx.models.dao;
+package mx.edu.utez.warehousemanagerfx.models;
 
-import mx.edu.utez.warehousemanagerfx.models.Branch;
 import mx.edu.utez.warehousemanagerfx.utils.database.OracleDatabaseConnectionManager;
 
 import java.sql.Connection;
